fix: use 32 bit instead of 8 bit for xp-related variables

+++ applications/services/dolphin/helpers/dolphin_state.c
@@ -112,7 +112,7 @@ uint32_t dolphin_state_xp_above_last_levelup(uint32_t icounter) {
     }
 
     uint8_t level = dolphin_get_level(icounter);
-    uint8_t level_threshold = REQUIRED_XP(level);
+    uint32_t level_threshold = REQUIRED_XP(level);
     return icounter - level_threshold;
 }
 
@@ -123,7 +123,7 @@ uint32_t dolphin_state_xp_to_levelup(uint32_t icounter) {
 
     uint8_t level = dolphin_get_level(icounter);
     uint8_t next_level = level + 1;
-    uint8_t next_level_threshold = REQUIRED_XP(next_level);
+    uint32_t next_level_threshold = REQUIRED_XP(next_level);
     return next_level_threshold - icounter;
 }
